LNURL-pay: move non-E2E tests to sdk-common
