+++ libs/sdk-common/src/lnurl/mod.rs
@@ -34,3 +34,13 @@ pub(crate) fn maybe_replace_host_with_mockito_test_host(
     // When not called from a test, we fallback to keeping the URL intact
     Ok(lnurl_endpoint)
 }
+
+#[cfg(test)]
+mod tests {
+    use bitcoin::secp256k1::rand;
+    use bitcoin::secp256k1::rand::distributions::{Alphanumeric, DistString};
+
+    pub fn rand_string(len: usize) -> String {
+        Alphanumeric.sample_string(&mut rand::thread_rng(), len)
+    }
+}
