+++ libs/sdk-common/src/lnurl/specs/pay.rs
@@ -451,3 +451,202 @@ pub mod model {
         }
     }
 }
+
+#[cfg(test)]
+pub(crate) mod tests {
+    use aes::cipher::{block_padding::Pkcs7, BlockDecryptMut, BlockEncryptMut, KeyIvInit};
+    use anyhow::Result;
+    use bitcoin::hashes::{sha256, Hash};
+
+    use crate::lnurl::specs::pay::*;
+    use crate::lnurl::tests::rand_string;
+
+    fn get_test_pay_req_data(
+        min_sendable: u64,
+        max_sendable: u64,
+        comment_len: u16,
+    ) -> LnUrlPayRequestData {
+        LnUrlPayRequestData {
+            min_sendable,
+            max_sendable,
+            comment_allowed: comment_len,
+            metadata_str: "".into(),
+            callback: "http://localhost:8080/callback".into(),
+            domain: "localhost".into(),
+            allows_nostr: false,
+            nostr_pubkey: None,
+            ln_address: None,
+        }
+    }
+
+    #[test]
+    fn test_lnurl_pay_validate_input() -> Result<()> {
+        assert!(validate_user_input(100_000, &None, 0, 100_000, 0).is_ok());
+        assert!(validate_user_input(100_000, &Some("test".into()), 0, 100_000, 5).is_ok());
+
+        assert!(validate_user_input(5000, &None, 10_000, 100_000, 5).is_err());
+        assert!(validate_user_input(200_000, &None, 10_000, 100_000, 5).is_err());
+        assert!(validate_user_input(100_000, &Some("test".into()), 10_000, 100_000, 0).is_err());
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_lnurl_pay_validate_success_action_encrypt_decrypt() -> Result<()> {
+        // Simulate a preimage, which will be the AES key
+        let key = sha256::Hash::hash(&[0x42; 16]);
+        let key_bytes = key.as_inner();
+
+        let iv_bytes = [0x24; 16]; // 16 bytes = 24 chars
+        let iv_base64 = base64::encode(iv_bytes); // JCQkJCQkJCQkJCQkJCQkJA==
+
+        let plaintext = "hello world! this is my plaintext.";
+        let plaintext_bytes = plaintext.as_bytes();
+
+        // hex = 91239ab5d94369a18474ee58372c7d0fcee5e227903f671bfe19ef32f1cada804d10f0f006265289d936317343dbc0ca
+        // base64 = kSOatdlDaaGEdO5YNyx9D87l4ieQP2cb/hnvMvHK2oBNEPDwBiZSidk2MXND28DK
+        let ciphertext_bytes =
+            &hex::decode("91239ab5d94369a18474ee58372c7d0fcee5e227903f671bfe19ef32f1cada804d10f0f006265289d936317343dbc0ca")?;
+        let ciphertext_base64 = base64::encode(ciphertext_bytes);
+
+        // Encrypt raw (which returns raw bytes)
+        let res = Aes256CbcEnc::new_from_slices(key_bytes, &iv_bytes)?
+            .encrypt_padded_vec_mut::<Pkcs7>(plaintext_bytes);
+        assert_eq!(res[..], ciphertext_bytes[..]);
+
+        // Decrypt raw (which returns raw bytes)
+        let res = Aes256CbcDec::new_from_slices(key_bytes, &iv_bytes)?
+            .decrypt_padded_vec_mut::<Pkcs7>(&res)?;
+        assert_eq!(res[..], plaintext_bytes[..]);
+
+        // Encrypt via AesSuccessActionData helper method (which returns a base64 representation of the bytes)
+        let res = AesSuccessActionData::encrypt(key_bytes, &iv_bytes, plaintext.into())?;
+        assert_eq!(res, base64::encode(ciphertext_bytes));
+
+        // Decrypt via AesSuccessActionData instance method (which returns an UTF-8 string of the plaintext bytes)
+        let res = AesSuccessActionData {
+            description: "Test AES successData description".into(),
+            ciphertext: ciphertext_base64,
+            iv: iv_base64,
+        }
+        .decrypt(key_bytes)?;
+        assert_eq!(res.as_bytes(), plaintext_bytes);
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_lnurl_pay_validate_success_action_aes() -> Result<()> {
+        assert!(AesSuccessActionData {
+            description: "Test AES successData description".into(),
+            ciphertext: "kSOatdlDaaGEdO5YNyx9D87l4ieQP2cb/hnvMvHK2oBNEPDwBiZSidk2MXND28DK".into(),
+            iv: base64::encode([0xa; 16])
+        }
+        .validate()
+        .is_ok());
+
+        // Description longer than 144 chars
+        assert!(AesSuccessActionData {
+            description: rand_string(150),
+            ciphertext: "kSOatdlDaaGEdO5YNyx9D87l4ieQP2cb/hnvMvHK2oBNEPDwBiZSidk2MXND28DK".into(),
+            iv: base64::encode([0xa; 16])
+        }
+        .validate()
+        .is_err());
+
+        // IV size below 16 bytes (24 chars)
+        assert!(AesSuccessActionData {
+            description: "Test AES successData description".into(),
+            ciphertext: "kSOatdlDaaGEdO5YNyx9D87l4ieQP2cb/hnvMvHK2oBNEPDwBiZSidk2MXND28DK".into(),
+            iv: base64::encode([0xa; 10])
+        }
+        .validate()
+        .is_err());
+
+        // IV size above 16 bytes (24 chars)
+        assert!(AesSuccessActionData {
+            description: "Test AES successData description".into(),
+            ciphertext: "kSOatdlDaaGEdO5YNyx9D87l4ieQP2cb/hnvMvHK2oBNEPDwBiZSidk2MXND28DK".into(),
+            iv: base64::encode([0xa; 20])
+        }
+        .validate()
+        .is_err());
+
+        // IV is not base64 encoded (but fits length of 24 chars)
+        assert!(AesSuccessActionData {
+            description: "Test AES successData description".into(),
+            ciphertext: "kSOatdlDaaGEdO5YNyx9D87l4ieQP2cb/hnvMvHK2oBNEPDwBiZSidk2MXND28DK".into(),
+            iv: ",".repeat(24)
+        }
+        .validate()
+        .is_err());
+
+        // Ciphertext is not base64 encoded
+        assert!(AesSuccessActionData {
+            description: "Test AES successData description".into(),
+            ciphertext: ",".repeat(96),
+            iv: base64::encode([0xa; 16])
+        }
+        .validate()
+        .is_err());
+
+        // Ciphertext longer than 4KB
+        assert!(AesSuccessActionData {
+            description: "Test AES successData description".into(),
+            ciphertext: base64::encode(rand_string(5000)),
+            iv: base64::encode([0xa; 16])
+        }
+        .validate()
+        .is_err());
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_lnurl_pay_validate_success_action_msg() -> Result<()> {
+        assert!(MessageSuccessActionData {
+            message: "short msg".into()
+        }
+        .validate()
+        .is_ok());
+
+        // Too long message
+        assert!(MessageSuccessActionData {
+            message: rand_string(150)
+        }
+        .validate()
+        .is_err());
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_lnurl_pay_validate_success_url() -> Result<()> {
+        let pay_req_data = get_test_pay_req_data(0, 100_000, 100);
+
+        assert!(UrlSuccessActionData {
+            description: "short msg".into(),
+            url: pay_req_data.callback.clone()
+        }
+        .validate(&pay_req_data)
+        .is_ok());
+
+        // Too long description
+        assert!(UrlSuccessActionData {
+            description: rand_string(150),
+            url: pay_req_data.callback.clone()
+        }
+        .validate(&pay_req_data)
+        .is_err());
+
+        // Different Success Action domain than in the callback URL
+        assert!(UrlSuccessActionData {
+            description: "short msg".into(),
+            url: "https://new-domain.com/test-url".into()
+        }
+        .validate(&pay_req_data)
+        .is_err());
+
+        Ok(())
+    }
+}
